Working prototype for the character note edit.
Uses a React component, to be renamed to CharacterNote.  The
window.fetch code is very clumsy and needs attention.  Tests are broken,
as I've not yet found the correct way to mock window.fetch.  Robot test
is imcomplete and broken.

But the notes really can be updated fast!  This gives a real boost to
motivation to continue on making the sheet an app rendered in the
browser.