Enable modification for existing skills.
Currently implemented by passing the modification event all up to
StatBlock which does the actual rest.patch() call;  should perhaps be
changed to do the REST call in the child component and on success pass
the modification up.  This would allow disabling the control while the
operation is ongoing, to prevent spurious errors and cases where React
and backend would be in different states.